+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/IrBindingGraph.kt
@@ -493,14 +493,15 @@ internal class IrBindingGraph(
       with(metroContext) { reportError(message, location) }
     }
 
-    val dependents = reverseAdjacency[binding.typeKey] ?: return
-    for (dependentKey in dependents) {
-      val dependentBinding = bindings[dependentKey] ?: continue
-      if (dependentBinding !is Binding.Assisted) {
-        reportInvalidBinding(
-          dependentBinding.parametersByKey[binding.typeKey]?.location
-            ?: dependentBinding.reportableLocation
-        )
+    reverseAdjacency[binding.typeKey]?.let { dependents ->
+      for (dependentKey in dependents) {
+        val dependentBinding = bindings[dependentKey] ?: continue
+        if (dependentBinding !is Binding.Assisted) {
+          reportInvalidBinding(
+            dependentBinding.parametersByKey[binding.typeKey]?.location
+              ?: dependentBinding.reportableLocation
+          )
+        }
       }
     }
     roots[binding.typeKey]?.let { reportInvalidBinding(it.declaration?.locationOrNull()) }