+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/transformers/AssistedFactoryTransformerTest.kt
@@ -1022,4 +1022,42 @@ class AssistedFactoryTransformerTest : MetroCompilerTest() {
       )
     }
   }
+
+  @Test
+  fun `assisted types cannot be depended on directly - accessor with no other ref`() {
+    compile(
+      source(
+        """
+            class ExampleClass @Inject constructor(
+              @Assisted val count: Int,
+              val message: String,
+            )
+
+            @AssistedFactory
+            interface ExampleClassFactory {
+              fun create(count: Int): ExampleClass
+            }
+
+            @DependencyGraph
+            interface ExampleGraph {
+              // The omission of ExampleClassFactory from accessors is intentional, prevents
+              // regression of https://github.com/ZacSweers/metro/issues/538 caused by existence
+              // of other dependents short-circuiting the check on roots
+              val exampleClass: ExampleClass
+
+              @Provides val string: String get() = "Hello, world!"
+            }
+          """
+          .trimIndent()
+      ),
+      expectedExitCode = COMPILATION_ERROR,
+    ) {
+      assertDiagnostics(
+        """
+          e: ExampleClass.kt:21:3 [Metro/InvalidBinding] 'test.ExampleClass' uses assisted injection and cannot be injected directly. You must inject a corresponding @AssistedFactory type instead.
+        """
+          .trimIndent()
+      )
+    }
+  }
 }